+++ src/helpers/define-api-host.ts
@@ -1,5 +1,10 @@
 import { EChains } from '../types'
-import { polygonApiUrl, baseApiUrl, mumbaiApiUrl } from '../configs'
+import {
+  polygonApiUrl,
+  baseApiUrl,
+  mumbaiApiUrl,
+  baseGoeliApiUrl
+} from '../configs'
 
 type TDefineApiHost = (
   chainId: number | null
@@ -10,6 +15,8 @@ const defineApiHost: TDefineApiHost = (chainId) => {
     case EChains.polygon: return polygonApiUrl
     case EChains.base: return baseApiUrl
     case EChains.mumbai: return mumbaiApiUrl
+    case EChains.baseGoerli: return   baseGoeliApiUrl
+
     default:
       return null
   }